backend/requirements.txt

django==5.0.5
djangorestframework==3.14.0
django-cors-headers==4.3.1
djangorestframework-simplejwt==5.3.1
Pillow==10.2.0
python-dotenv==1.0.1
# Uncomment for SMS functionality with Twilio
# twilio==8.0.0
